Moved operation classes
