+++ product_code/operation_service/impl/src/main/scala/de/upb/cs/uc4/operation/impl/actor/OperationDataList.scala
@@ -1,6 +1,6 @@
 package de.upb.cs.uc4.operation.impl.actor
 
-import de.upb.cs.uc4.shared.client.operation.OperationData
+import de.upb.cs.uc4.hyperledger.api.model.operation.OperationData
 import play.api.libs.json.{ Format, Json }
 
 case class OperationDataList(operationDataList: Seq[OperationData])
